Added Player Hit Box
Hit Box added to the Player swoed, so I can start to implement an attack routine for the player.

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public abstract class Enemy : MonoBehaviour
 {
     [SerializeField] protected int health;   //protected = only object that ihnerit can modify it.
-    [SerializeField] protected int speed;    //protected = only object that ihnerit can modify it.
+    [SerializeField] protected float speed;    //protected = only object that ihnerit can modify it.
     [SerializeField] protected int gems;     //protected = only object that ihnerit can modify it.
 
     [SerializeField] protected Transform pointA, pointB; //waypoints
